requirements.txt
django==4.1
djangorestframework==3.15.1
markdown==3.6
drf-yasg==1.21.7